Refactor: Transition project to MySQL with EF Core migrations
This commit encompasses the transition from SQL Server to MySQL, addressing compatibility and configuration adjustments necessary for the switch. It includes updates to the DbContext and connection strings to align with MySQL's requirements, along with modifications to the EF Core migrations to ensure compatibility with the MySQL database schema. Additional updates include adjustments to environment-specific settings to leverage environment variables for database connectivity and other configurations in production scenarios.

Key changes:
- Updated `DbContext` for MySQL compatibility.
- Revised EF Core migrations for MySQL database schema.
- Altered connection strings in `appsettings.json` files for MySQL.
- Implemented environment variable usage for sensitive and environment-specific configurations.
- Resolved startup configuration issues to correctly identify the development environment and configure services accordingly.

This transition enhances the project's flexibility in database management and deployment, aligning with the cross-platform development goals.
